backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
faiss-cpu
sentence-transformers
tqdm
requests
rank-bm25